rm lockfile, move async stuff into nonblocking
